refactor metar class, bug fixes
