+++ MetarSharp/Metar/RecentWeather.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MetarSharp
+{
+    public class RecentWeather
+    {
+        public string RecentWeatherRaw { get; set; }
+
+        public string RecentWeatherTypeRaw { get; set; }
+
+        public string RecentWeatherDecoded { get; set; }
+    }
+    
+}
