+++ MetarSharp/Metar/RunwayCondition.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MetarSharp
+{
+    public enum Condition
+    {
+        ClearAndDry,
+        Damp,
+        Wet,
+        Frost,
+        DrySnow,
+        WetSnow,
+        Slush,
+        Ice,
+        CompactedSnow,
+        FrozenRidges,
+        Unknown
+    }
+    public class RunwayCondition
+    {
+        public string? RunwayConditionRaw { get; set; }
+        public Condition Condition { get; set; }
+    }
+}
