+++ MetarSharp/ParseMetar/Visibility/ParseFromKilometer.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using static MetarSharp.Parse.ParseVisibility;
 
 namespace MetarSharp.Parse
 {
@@ -11,7 +12,34 @@ namespace MetarSharp.Parse
     {
         internal static Visibility ParseVisibility(GroupCollection groups)
         {
-            return new Visibility();
+            Visibility visibility = new();
+
+            #region STANDARD
+
+            visibility.VisibilityRaw = groups[7].Value;
+
+            visibility.IsVisibilityMeasurable = true;
+
+            visibility.ReportedVisibility = double.Parse(groups[8].Value);
+
+            visibility.VisibilityUnit = VisibilityUnit.Kilometers;
+            visibility.VisibilityUnitRaw = "KM";
+            visibility.VisibilityUnitDecoded = "Kilometers";
+
+            #endregion
+
+            if (groups[13].Success)
+            {
+                visibility.HasVisibilityLowestValue = true;
+                visibility.LowestVisibility = double.Parse(groups[14].Value);
+                (
+                    visibility.LowestVisibilityDirection,
+                    visibility.LowestVisibilityDirectionDecoded
+                ) = GetCardinalDirection(groups[15].Value);
+                visibility.LowestVisibilityDirectionRaw = groups[15].Value;
+            }
+
+            return visibility;
         }
     }
 }
