+++ MetarSharpDebugger/Program.cs
@@ -25,12 +25,20 @@ internal class Program
         timer.Start();
 
         //You can enter your metars here
-        Metar metar = ParseMetar.ParseFromString("KCOU 182054Z 18011KT 10SM 0050E VCTS FEW031 FEW055 BKN120 25/22 A2995 WS R25 RESN BLU RMK HALLO");
+
+        var lines = File.ReadAllLines("../Metars.txt");
+        List<Metar> metars = new();
+
+        foreach(var line in lines)
+        {
+            Metar metar = ParseMetar.ParseFromString(line);
+            metars.Add(metar);
+        }
 
         ///Just for diagnostics/to check execution time 
         timer.Stop();
         var executeTime = timer.ElapsedMilliseconds;
-        var timerPerMetar = Math.Round((double)executeTime / m.Count, 5);
+        var timerPerMetar = Math.Round((double)executeTime / metars.Count, 5);
     }
 }
